faq proxy refactored
